requirements.txt
faiss-cpu
google-generativeai
jupyter
matplotlib
pandas
plotly
seaborn
sentence-transformers
streamlit
torch
transformers